requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.1.3
django-cors-headers==3.13.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
phonenumbers==8.12.50
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
simplejson==3.17.6
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
